+++ tools/websockets_test.py
@@ -34,4 +34,4 @@ async def main():
             else:
                 print(f"< {zlib.decompress(response, 16 + zlib.MAX_WBITS)}")
 
-asyncio.get_event_loop().run_until_complete(main())
+asyncio.run(main())